* Move libcli/ to tdns/cli/. Fixes #94.
