+++ tdns/cli/ping.go
